// src/keymaps/keymap.interface.ts
export interface KeyMapInterface {
  ENTER: number;
  LEFT: number;
  RIGHT: number;
  UP: number;
  DOWN: number;
  BACK: number;
  DIGIT_1: any;
  DIGIT_2: any;
  DIGIT_3: any;
  DIGIT_4: any;
  DIGIT_5: any;
  DIGIT_6: any;
  DIGIT_7: any;
  DIGIT_8: any;
  DIGIT_9: any;
  DIGIT_0: any;
  STOP?: any;
  PLAY?: any;
  PAUSE?: any;
  PLAY_PAUSE?: any;
  FWD?: any;
  RWD?: any;
}